// src/app/core/models/my-contacts.model.ts
export class MyContacts {
      uuid: string;
      id: string;
      first_name: string;
      phone_dial_code: string;
      phone_number: string;
      email: string;
      advisor_id_number: number;
      advisor_role: string;
}


export class CustomerInforAdvisor {
      id: string;
      first_name: string;
      phone_dial_code: string;
      phone_number: string;
      email: string;
      address_line1: string;
      address_line2: string;
      address_line3: string;
      postal_code: string;
      state_code: string;
      country_code: string;
      customer_id : number;
}

export class MyCustomer {
      mAddress: string;
      mCreated: Date;
      mDisplayName: string;
      mEmail1: string;
      mEmail2: string;
      mId: number;
      mIsReceipt: number;
      mModified: Date;
      mRepresentative1: string;
      mRepresentative2: string;
      mStatus: string;
      mTelNumber1: string;
      mTelNumber2: string;
}